Fix reviewer issues: fixtures to conftest, extract helpers, fix assertions
- Move fixtures to conftest.py per project convention
- Extract _make_snapshot() and _setup_host_with_sandbox() helpers to
  reduce duplication across tag/snapshot tests
- Move make_testing_modal_interface and make_testing_provider to a
  testing.py module for shared use between conftest and tests
- Fix vacuous assertions (assert len(x) >= 0) with meaningful checks
- Add logging to deploy() error handler in TestingModalInterface
- Remove test that only tested the test helper function
